New way to create division problem.
Use randomly generated numbers to find the product. One of the numbers
then becomes the factor. Note that a new set of numbers are randomly
generated for division, and they are limited to be small enough such
that their product will obey the number of digits set by the user.